Add curved walls feature
An option in the user menu gives the user the option to have bricks which look like they have slightly concave walls just like in real life. This especially improves the realism when there are reflections in the build. However, the option results in an increase in render time (about + 25% on my machine).
The look of the curved walls is achieved with standard Cycles group node consisting of only standard nodes that connects into the normal input of the Lego material to give maximum compatibility. The curved look is not implemented in instruction rendering mode as it would increase the rendering times without serving a function.