Add grainy texture to slopes
Adds a rough, grainy texture to the sloping faces of certain bricks as is the case in real life. It's not possible to find which faces should receive that texture and which ones do not in a fully automatically way, so that's why the script uses a hard coded dictionary which should be updated as new slope bricks with a grainy texture are designed or when it appears that the provided dictionary is not complete yet. The dictionary contains the part number and the cotan of the slopes which should receive the texture. It could also be possible to include the hard coded data in a separate file instead of embedded in the code so people can access it more easily to update it themselves.

Determination of which faces should receive the texture happens as follows:
1) The part has to appear in the provided dictionary
2) The face has to point outwards relative to the part origin. The default LDraw origin does not lie at the inside for every part, so another hard-coded dictionary is provided to displace the origin to separate the inner from the outer sloped faces.
3) The slope of the face has to correspond to the cotan provided in the hard-coded dictionary. The provided value is for a projection of the slope on the local horizontal x or z axis which are easier to define accurately than the angle between the face normal and the horizontal plane in the case of sloped wedges. Some bricks (the new Nexoknight pointy slopes) have compound curved faces with a grainy texture, making it difficult to define the slope for every single face in the .dat file. For that purpose, it is also possible to define an interval in the dictionary rather than a single value for the cotan.
All in all, not a lot of hard-coded data is necessary to accurately assign the slope material, even for parts with a high polygon count like decorated slopes.

The grainy material is achieved with a voronoi texture and a bump node. The texture uses the object coordinates as vector input in order to get non-distorted bumps on all parts with a consistent scale. The disadvantage is that one importing multiple models into the same .blend file with different scales, the material is always the same so the absolute scale of the bumps will be the same to the relative size will be off, but this is a rare case.