Fix use of periods in user interface strings
Tooltip strings for check boxes should not end with a period, as
Blender will automatically add one.

Tooltip strings for individual items in a multi-select should,
because Blender won't add any here.

This patch removes and adds periods as appropriate so that every
string ends up with exactly one period when displayed in the UI.

Fixes #27